app/requirements.txt
fastapi
uvicorn
pymongo
pandas
pydantic
python-dotenv
pydantic-settings
python-multipart